requirements.txt
torch>=2.1
transformers>=4.40
accelerate>=0.27
bitsandbytes>=0.42
peft>=0.10
datasets>=2.19
sentence-transformers>=2.7
faiss-cpu>=1.7
arxiv==1.4.8
tqdm>=4.66 